refactor to one method
