+++ x/vm/client/cli/query.go
@@ -2,7 +2,7 @@ package cli
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	common "github.com/cosmos/evm/precompiles/common"
+	"github.com/cosmos/evm/utils"
 	"github.com/spf13/cobra"
 
 	rpctypes "github.com/cosmos/evm/rpc/types"
@@ -213,7 +213,7 @@ func HexToBech32Cmd() *cobra.Command {
 		Example: "evmd query evm 0x-to-bech32 0x7cB61D4117AE31a12E393a1Cfa3BaC666481D02E",
 		Args:    cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			cmd.Println(common.Bech32FromHexAddress(args[0]))
+			cmd.Println(utils.EthHexToCosmosAddr(args[0]))
 			return nil
 		},
 	}
@@ -227,15 +227,15 @@ func Bech32ToHexCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:     "bech32-to-0x",
 		Short:   "Get the bech32 address for a given 0x address",
-		Long:    "Get the evm 0x address for a given bech 32 address.",
-		Example: "evmd query evm 0x-to-bech32 ",
+		Long:    "Get the bech32 address for a given 0x address.",
+		Example: "evmd query evm bech32-to-0x cosmos10jmp6sgh4cc6zt3e8gw05wavvejgr5pwsjskvv",
 		Args:    cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			addr, err := sdk.AccAddressFromBech32(args[0])
 			if err != nil {
 				return err
 			}
-			hex, err := common.HexAddressFromBech32String(addr.String())
+			hex, err := utils.HexAddressFromBech32String(addr.String())
 			cmd.Println(hex.String())
 			return nil
 		},
